pyproject.toml
[tool.poetry]
name = "order-stock-market"
version = "0.1.0"
description = ""
authors = ["Felipe Carvalho <felipe.carvalho07@gmail.com>"]

[tool.black]
line-length = 91
target-version = ["py38"]

[tool.isort]
line_length = 91
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
force_grid_wrap = 0
known_local_folder="orders"
sections="FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
default_section="THIRDPARTY"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "stocks.settings"
addopts = "-v -x --cov=orders --cov-report=term-missing --ignore=stocks"
testpaths = [
    "orders",
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
prettyconf = "^2.2.1"
uuid = "^1.30"
django-cors-headers = "^3.7.0"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.7b0"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-django = "^4.4.0"
drf-yasg = "^1.20.0"
isort = "^5.9.3"
freezegun = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
